JS Code to Fix SharePoint Document Library
This Code can be used to fix Document Libraries in SharePoint 2013 in order to wrap the Application Icon and the Title of an entry with the link of the specific file. So you can now also download or open the file via the title-text or the icon at the beginning of each entry.